Cleanup, pre-cherry-pick
Some changes that I partially implemented a while ago as part of a
larger commit really should be in their own separate commit.  The reason
for this is that I want to cherry-pick part of the changes but not all
of them.

This commit un-does the changes I want to cherry pick.  They will be
re-added, and properly finished in another commit prior to
cherry-picking.