Improve Memory Mgt. and Performance of RF Params Ops
This commit allows Xyce to determine which RF parameter types (S, Y
or Z) have been requested by a .LIN or .PRINT AC line.  This helps with
memory management and performance since the Sparams_ and Zparams_
matrices in the AC object don't have to be resized,  or converted
from the YParams_ matrix, if the netlist doesn't need S or Z
parameters.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199