Reformat latex input for measure section
Most of this section was written with very long lines, and many
"\item" sections were not separated from the previous one by any white
space.  This makes absolutely no difference to LaTeX, but can make it
difficult to work with in an 80-column editor window (the Xyce style,
remember, is to format all C++ code to 80-column lines, so a number of
us are using editor windows of that size, or that have some manner of
showing where lines should break).

I also took the opportunity to remove a lot of what Git calls
"whitespace errors."  (these are blank characters at the ends of lines)

No actual change to the formatting by latex results from these edits.