Initial Capability for Embedded Sampling Outputters
This is the initial capability for the outputters for embedded
sampling (.EMBEDDEDSAMPLING).  Highlights of this new capability
are as follows.  For the new .PRINT ES line, which is only compatible
with TRAN and DC analyses are:

  1) FORMAT=<STD|NOINDEX|GNUPLOT|SPLOT> will make output files of the
form <netlistName>.ES.prn.

  2) FORMAT=CSV will make output files of the form <netlistName>.ES.csv.

  3) FORMAT=TECPLOT will make output files of the form <netlistName>.ES.dat.

  4) FORMAT=<RAW|PROBE|TOUCHSTONE|TOUCHSTONE2> will default to FORMAT=STD.

  5) The column list in each output file matches those that are generated
by the function EmbeddedSampling::hackEnsembleOutput().

  6) A request for -r or -a, when .EMBEDDEDSAMPLING is used in the netlist,
is currently a parsing error.

  7) For a -o <fileName> command line option, the output from the .PRINT ES
line will be sent to <fileName> in FORMAT=STD.

  8) The output from the .PRINT TRAN or .PRINT DC lines in the netlist
will occur, if -o is not specified on the command line.  (Note: this may
be changed in the future.)

  9) New .PRINT parameters (outputSampleStats and outputAllSamples) that
are specific to the .PRINT ES line were added.  Their default values are
TRUE and FALSE.  These parameters should be silently ignored by other
.PRINT lines.

SON Bug 1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201