Stop eliminating constant subexpressions
PTcheck_ is a recursive method of the expression class that tries to
perform common subexpression elimination.

This function is absolutely a bottleneck for large tables such as
those that result from using the "tablefile" feature.  Those tables
are typically huge, and PTcheck_ is basically an N^2 operation.

This commit makes PTcheck_ skip trying to do common subexpression
elimination if the subexpression in question is just a constant.  The
benefit of doing "common constant" elimination is just to save the
memory of a few single ExpressionNode objects, and is not worthy of
the effort it takes to find them in a big table --- and even for a big
expression it's probably not that big of a savings.  For a tablefile,
the odds that there are duplicate constants all over the table is low,
and so performing an N^2 search to come up with the conclusion that
there aren't any is insane.

I have also discovered that Xyce is, for some reason, performing a
copy of an expression at least once after the expression is parsed ---
I have not tracked down why.  Unfortunately, because of the way that
the copy is performed, it is always necessary to call "PTcheck_" on
the copy.  That's because the copy is performed somewhat inefficiently
and will REINTRODUCE into the copy any common subexpressions that were
eliminated when the original expression was first parsed!

One could seriously improve the performance of the expression package
by fixing either how the copy is done, or eliminating the need for a
copy in the first place.

This commit does not completely fix bug 1176, but seriously improves
the performance on large tables, so I am committing it with
confidence.  I cannot close the bug based on this one commit, though.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1176