Fix Multi-Parameter Touchstone Output from .LIN Lines
The previous commit (82a1e65c), intended to improve memory management
and performance of RF parameter ops such as S(), Y() and Z(), had
the unwanted side effect of making Xyce only produce correct Touchstone
output from the last .LIN line.  An example is:

  .LIN FORMAT=TOUCHSTONE FILE=sparams1.cir.sparams.s1p
  .LIN FORMAT=TOUCHSTONE LINTYPE=y FILE=sparams1.cir.yparams.s1p

In that case, the file sparams1.cir.sparams.s1p would only contain
the header line and the Freq column.  It would not have any of the
S-parameter data, since the Sparams_ matrix was not being made.  This
was because only one ACLIN option block was being processed by the AC
object.  This  commit makes it so that all of the ACLIN option blocks
get processed. It also makes it so that a .LIN analysis is done if
SPARCALC=1 on any of the .LIN lines.

There are also a lot of "editorial changes" of AC to ACLin in function
and variable names.  This makes it clearer that these functions are
about handling for an ACLIN option block, generated from a .LIN line,
and not for an AC option block.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199