Fix so that resampling will work in parallel
So far we have not been testing any of the PCE code in parallel
regression testing.  Today, we attempted to do this, but in the first
run several tests failed.  Some of them failed becuase the resampling
functions were hanging and then causing subsequent memory errors.

The reason for this is that the resampling is only performed on proc
zero.  This is fine, but it needs a random number seed so it calls the
function UQ::getTheSeed.  That would also be fine, except that this
function uses a comm object and has safe barriers inside of it.  So, if
it is only called on proc 0 then there will be problems.

This commit moves the call to UQ::getTheSeed to a point in the function
that is outside of the proc==0 if-statement.  Now this problem is
resolved.