Add template to propagate dependencies to AF outputs
ADMS applies a series of "implicit" templates that, among other
things, traverse the ADMS data tree and try to propagate probe and
variable dependencies upward.  Unfortunately, there is a huge gap in
this, notably that these dependencies, when they are in an assignment,
only get propagated to the top-level of the RHS expression and the
LHS.

If the RHS contains analog function calls, the expression involving
the call gets all of the dependencies of the function's arguments, as
does the LHS.

But none of these are propagated to the OUTPUT variables of the analog
function, if there are any.  This is a huge problem for MVSG-HV (and
so far, ONLY MVSG-HV), because they use tons of these.  I kludged
around it by hacking the MVSG-HV source code to create a series of
assignment statements that fake out ADMS, but it was an ugly mess.

This commit adds a template (so far unused except in the
"html_params.xml" debugging back-end) that can be used to scan the
entire data tree for all analog function calls and then carefully
propagate all dependencies that the input arguments have into the
output variables.

It is a messy operation, because the actual arguments may be
expressions in their own right, and because the implicit rules do NOT
propagate dependencies except to the top level expression we have to
do it all over again.

I know that my templates are not complete --- they will do quite the
wrong thing if one puts an analog function call with output variables
in the actual arguments of another analog function call with output
variables.  That was just too many levels of recursion for me to wrap
my brain around.  It should still be correct for straight analog
function calls with NO outputs other than the return value.

The next step is to make templates that add output variables of analog
function calls to the "@assignedVars" arrays that are used to make
sure variables get declared.  Then I can remove some kludges that have
been created for MVSG-HV.

This work is absolutely a prerequisite for finishing the bug 1168
work, but I am doing it on a separate branch because the primary model
that really needs this isn't yet usable with the new back-end, and I
want to merge this into develop long before the new back-end is ready.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168

Remove text that confuses Emacs

Emacs' "xml-mode" hates when I put multiple dashes in comments, and
its XML validator displays a little "Invalid" notice in the little
info bar at the bottom.  Since the validator is really useful to help
me find mistakes in XML markup as I write it, it is best if it says
"Invalid" only when there's an actual mistake.

It is really hard to find things like this, as while Emacs will
helpfully tell me that the buffer contains invalid XML, it won't
actually point me at the spot that's upsetting it.