Put ostream<< operator inside Xyce namespace
When I try to compile Xyce with Amesos2 using clang (instead of gcc), the
compilation fails in one of the Basker headers.  It fails because of the
insertion operattor (<<) for the genericBlockMatrixEntry object not
being in the Xyce namespace.

Gcc didn't seem to care about this, but clang does.

As it turns out, it was really easy to move the insertion operator from
outside the Xyce namespace to inside.  All I had to do was move a
right-hand curly brace from one part of the file to another part roughly
10 lines later.  Problem solved.