Recognize output vars anywhere in list
My earlier commit (which was cribbed off of some experimental work in
the Qucs/ADMS repo) improperly assumed that all analog function output
variables were the first ones in the argument list.

This assumption snuck in by using a select expression  that created a
list ONLY of output variables in the function definition, then
assuming that the position of a variable in this list corresponded to
its position in the actual argument list.  This works only when all
output variables are at the beginning.  If they are interleaved with
input variables, it's completely wrong.

Now we select ALL variables in the definition, but then when we loop
over them, do nothing when they're not input variables.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168