Correct analog function kludge code in MVSG-HV
To make MVSG-HV work with the Xyce/ADMS version prior to my current
hackage, I had to add a block of code that worked around the failure
to recognize output variables of analog functions and failure to
propagate dependencies into them.  I did this by adding a block of
code that explicitly assigns to those variables a simple addition of
exactly the same probes that the function return value depends
on (which ADMS gets right).

Unfortunately, I did this with an early version of MVSG-HV, and when
we got the MVSG-HV version accepted by CMC I did NOT correctly redo
this section --- there were a number of changes that made the return
value depend on fewer probes.  The result was that now the hacked code
was adding too many probe dependencies, and then those were
propagating into the return value as well.  I discovered this during
my current hackage, when I was finding my fixed Xyce/ADMS (which DOES
propagate dependencies correctly to analog function output variables)
seemed to be missing some dependencies that the old code had.

Tracked it to the fact that the old workaround was just adding
unnecessary dependencies.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168