Get output variables of Analog Functions handled correctly
This commit makes Xyce/ADMS work around a serious deficiency of the
handling of analog functions by the ADMS implicit templates in
adms.implicit.xml (in the ADMS source code, and dumped into the
current working directory as ".adms.implicit.xml" whenever you run
admsXml).

The implicit templates of adms.implicit.xml scan the ADMS data tree
and propagate variable and probe dependencies in expressions around.
This is done in the template "e:dependency".  However, the following
things are not done:

  - Output variables of analog functions are, for all intents and
    purposes, assignments, and should get the same variable and probe
    dependencies that the return value of the function get.  They
    don't.

  - Because these variables don't get dependencies correctly
    propagated, the Jacobian datastructure ADMS produces (in the
    "adms.implicit.xml.module" template) is missing any jacobian
    entries that should arise from USING output variables of analog
    functions in contributions.

Furthermore, Xyce/ADMS was originally written in a way that imitated
how ngspice and Qucs use ADMS in several regards, not the least of
which is how we decide which variables to declare in C++ --- we look
through the ADMS data tree and declare anything that appears on the
LHS of an assignment.  But it is possible to use variables as output
variables of an analog function, thereby assigning to them without
them ever BEING on the LHS of an assignment.

A previous commit on this branch added a template to propagate
variable dependencies into output variables, and I explored using it
only with the html_params.xml "HTML back-end" that just outputs a file
of HTML with information about the module under consideration.

This commit modifies the Xyce/ADMS template "collectAssignedVariables"
so that it now can find variables that are only "assigned" by being
present in an analog function call whose dummy argument is marked as
an "output" variable.  It also modifies the
"fixAFOutputVariableDependencies" template created earlier so that it
also updates the jacobian data structure to add any jacobian entries
that would have been missed by ADMS's default operations.

With the code committed here, MVSG-HV now imports into Xyce and passes
the CMC test suite for that model just as well as it did with the old
back-end.

I have not yet tested it on all the other models that still require
the Sacado back-end (which at this point is only bsim6, bsimcmg 107,
FBH HBT_X, and MVSG-HV).  Since MVSG-HV is the only one of these that
HAS analog functions with output variables, I do not expect problems
with the others.  But there is always room to be surprised.

This work does NOT remove any Sacado usage, but it does remove an
obstacle to finishing the non-Sacado ADMS back-end for bug 1168.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168