Apply PCE analysis to .measure outputs
Earlier this year I adapted the sampling code to compute statistics on
.MEASURE outputs.  Originally I wrote the code to just do statistics on
solution variables and simple expressions that could be evaluated via
the Op classes.

However, as it turns out I never applied the PCE code to .MEASURE
outputs for some reason.  I hadn't noticed that until earlier this week.
This commit fixes that oversight.

Now, if the user specifies a list of measures on the .SAMPLING line, and
also specifies one of the PCE analyses to be run, that PCE analysis will
be applied to all the requested measures.