Add PCE coefficient output to hackEnsembleOutput
The hackEnsembleOutput function is likely to be obsolete very soon.

However, there was one last thing I meant to add to it that I never got
around to, and that was to make it possible to output PCE coefficients
to the plot file.

This commit adds that capability.  It also modifies some of the std
output logic.  There has been a stdOutput flag, which controlled most of
the output to the screen, except for the output of PCE coefficients,
which had their own flag.  Now, the PCE coeffient output to std output
is also controlled the std output flag.

The flag for outputting PCE coefficients still exists, but it is now
applied inside of the std output flag, and also inside of the
hackEnsembleOutput function.

I implemented this mostly to give some guidance to Pete for his
ES outputter work.

One complication is what names to give the PCE coefficients in the
output file.  I chose to make them very generic.  Possibly we can come
up with better names later.