Add modified ADMS "implicit" template file
ADMS runs an "implicit" set of XML (or rather ADMST) templates over a
Verilog file before executing any explicitly listed templates.  It
does this by dumping a copy of the file "adms.implicit.xml" from the
ADMS source tree into the local directory (as a hidden file named
".adms.implicit.xml") and processing it before doing anything else.

This behavior can be turned off by adding the "-x" flag to admsXml.
One would then have to provide an EXPLICIT file to run instead.

adms.implicit.xml does a lot of early processing of the ADMS data
tree, including propagating probe and variable dependencies all over
the place.  It fails to do so correctly for analog functions with
output variables in the argument list (as opposed to a return value).

This commit adds a copy of the default ADMS implicit templates file,
with the template "e:dependency" augmented to handle analog functions
better.

With this new "implicit" template file invoked explicitly, Xyce/ADMS
can now process MVSG-HV without the clumsy and incomplete
"fixAFOutputDependencies" hack I put in earlier on this branch.
Because it's done at the same time as all other dependency tracking,
there is no longer an error in the jacobian that needs to be fixed up,
and there is no longer a clumsy limitation on how analog function
output variables can be used downstream.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223