Small fix to parameter handling
I previously noticed and fixed this issue in the Sampling and
EmbeddedSampling classes.  Several of the PCE-related parameters were
being processed incorrectly.   They were being processed incorrectly in
that rather than reading the actual value from the netlist, they were
setting the relevant boolean to "true" if the parameter was being set at
all.

So, in other words, if the user set  OUTPUT_PCE_COEFS= true OR
false, the internal boolean variable outputPCECoeffs_ was being
unconditionally set to true.    This was of course, wrong behavior, but
easy to fix.

The fixes to embedded sampling and regular sampling were handled in a
different commit, but that commit also contained some other changes so
this particular change was inadvertantly not mentioned in the commit
message.