Remove MVSG-HV kludges and use new implicit templates
To work around the broken bits of the default ADMS implicit templates,
I had inserted really ugly and kludgy hacks to the MVSG-HV Verilog-A
source.

Now that I've got an explicitly-invoked set of fixed "implicit"
templates, this hack is no longer necessary.  I am removing it
altogether, and making the MVSG-HV Makefile invoke the new templates
explicitly when it builds this device.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223