Add regenerated MVSG-HV C++ code
During my work on bug 1168 and bug 1223, I discovered that the kludgy
hack that I'd inserted into MVSG-HV's Verilog-A source code to work
around ADMS's broken analog function handling was in fact slightly
wrong.

To work around the fact that something like this:
   foo = analogfunctioncall(bar,baz,input1,input2,input3)

fails to propagate the probe and variable dependence of input1,input2,
and input3 anywhere but into "foo" (i.e., not into the output
variables bar and baz), I'd added a block of clumsy code to a block of
code that would never actually run, but would be seen by ADMS and fix
the dependency problem.  Thing is, I did that with the FIRST version
of MVSG-HV we ever got, and when we got an updated version I just
copied it.  Turns out that the new version changed some function calls
and the dependencies were now different.  Fortunately, they were
different in a way that didn't actually break anything; some probes
went from being like "V(A,GND)" and "V(B,GND)" to "V(A,B)", thereby
making the actual code output work just the same, and in a few cases
they REMOVED dependencies and my code was inserting them again ---
again, this just bloated the C++ output with extra probes that weren't
needed, but didn't harm anything.

Once I fixed bug 1223 properly, this hack became unnecessary, AND
during testing of my fixes I realized that the significant differences
in the output C++ were in fact just indicative of my old hack being
WRONG.

So, I've removed that hack from the input verilog, and regenerated
this model using the fixed Xyce/ADMS that tracks things properly.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168