Copy bug 1223 fixes from develop
While working on bug 1223 and a particular related analog function
issue, I modified the xyceBasicTemplates.xml file on the develop
branch.  I've rebased my bug 1168 branch to develop so that I get
those changes, but now I have to back-port them into the non-sacado
version.

This has been tested by rebuilding all of the models that use the new
back-end and confirming that nothing changes except the creation date
in a comment.  Since that's all that's changed, I won't be committing
the regenerated code.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223