Make all models use new "implicit" templates
As part of fixing bug 1223, I created a copy of the default "implicit"
templates used by ADMS, but modified them to handle analog functions
better.  Now I'm changing all of the Makefiles for models for which we
use the experimental back-end so that they use these modified
templates explicitly, and don't use the implicit ones at all.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223