Make buildxyceplugin use new implicit templates
In my work on bug 1223, I created a modified version of the standard
ADMS file "adms.implicit.xml" that can be used explicitly as a
replacement for ADMS's version.  This version handles analog functions
with output arguments correctly, unlike ADMS's.

This commit makes this file be installed by "make install" for a
shared-library build of Xyce, and makes buildxyceplugin use it.

Now buildxyceplugin will properly process MVSG-HV without
Xyce-specific hacks on the input Verilog-A file.  It will now also
process other models that use analog functions in this way in the
future.

This commit also adds text to the release notes about the fix of this
bug.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223