Improvements to Embedded Sampling Outputters
This commit makes improvements to the outputters for Embedded
Sampling.  First, the output from .PRINT ES lines is now only made
when the .OPTIONS EMBEDDEDSAMPLES has a OUTPUTS= parameter on it.
This is how the EmbeddedSampling::hackEnsembleOutput() function
works.  Second, the conditional block that controls that output
was moved from EmbeddedSampling::hackEnsembleOutput() to the
function EmbeddedSampling::stepCallBack(), where it now "wraps"
the calls to both hackEnsembleOutput() and the function used for
.PRINT ES output, outputManagerAdapter_.outputEmbeddedSampling().
Third, outputEmbeddedSampling() is now called for non-Stokhos
builds.  Fourth, the string vectors of the PCE coefficients are now
made unconditionally on the first time that the outputters are
called.  This allows OUTPUT_PCE_COEFFS to be set differently on
the .PRINT ES and .OPTIONS EMBEDDEDSAMPLING lines.  Finally, a
common way of outputting the PCE coefficients is now used by both
outputEmbeddedSampling() and hackEnsembleOutput().  This should
improve maintainability until hackEnsembleOutput() is removed
from the code.

SON Bug 1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201