Checkpoint a few days' work
I have done a ton of typing this week and don't want to lose it in a
catastrophic disk fail.  Committing my work in progress, which is all
in one file and starts fleshing out templates that will be needed to
generate analog function derivative functions.  So far I've only done
the templates needed to expand expressions into their values and
derivatives with respect to analog function arguments.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168