Start emitting analog function derivatives
This very embryonic commit is able to emit functions that compute the
derivatives of analog functions with respect to a chosen variable.

An analog function such as:
   analog function foo
     input a,b;
     real a,b;

     begin
        <some computation that computes the value of foo>
        foo= <value>
     end
   endfunction

is already being translated into a general templated function that
requires no modification to get the value of foo.  However, to get the
derivatives of foo with respect to some variable "c", we also need a
function:
   d_foo(a,b,d_a_d_c,d_b_d_c)

where d_a_d_c is the derivative of a with respect to c and d_b_d_c is
the derivative of b with respect to c.

The d_foo function must compute foo just as normal, but in the process
also compute d_foo_d_a and d_foo_d_b.  At the end of the line, we
construct
   d_foo = d_foo_d_a*d_a_d_c + d_foo_d_b*d_b_d_c

The same thing must be done for ALL output variables (not just the
return value).

This commit accomplishes this for simple analog functions that do not
call other analog functions.

The reason for this limitation is that I have not yet written the part
of the templating that actually USES these functions, and until I do,
I won't be able to use them inside analog functions, either.

Unfortunately, MVSG-HV actually calls analog functions from within
other analog functions, so it remains the problem child that drives
all the difficult work on this branch.

In addition to the major functional change, this commit also fixes a
few errors in earlier commits and adds correct handling of the "log"
function (log base 10), which had previously been overlooked (because
no models that we work with use log, only ln).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168