Add support for "$vt" calls in analog function
$vt(temperature) is suppose to return CONSTKoverQ*temperature.

$vt is supposed to be the same thing, but evaluated at the current
global simulator temperature.

Use of $vt(temp) with an argument is a reasonable thing to do inside
an analog function, and FBH does it.  Use of $vt without an argument
will be unsupported in Xyce/ADMS, at least for now (because the analog
functions are just free functions in an appropriate namespace, not
member functions, and therefore can't get at anything "global").

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168