Improve Error Checking for S(), Y() and Z() Operators
This commit adds error checking for the S(), Y() and Z() operators
for the case where one of the indices is greater than the number of
ports in the netlist.  This approach is somewhat suboptimal since
the error checking is in the get() functions, rather than in the
struct RFparamsVariableOpBuilder class.  So, Xyce will error out
part way through making the output files, rather than during netlist
parsing.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199