Allow analog functions to call other AFs
This commit finally gets to the point where we can emit analog
function calls (with derivatives) inside of analog function derivative
definitions.  This was the big blocker to getting FBH and MVSG-HV
analog functions processed, because both of those models call their
own analog functions from within other analog functions.

In addition to that actual feature enhancement, this commit also does
a little code fixing and cleanup.

I had incorrect code generation for the derivative of a variable with
respect to another (it was emitting "variable_wrt" instead of
"d_variable_d_wrt").  It was putting a variable for the correct
derivative in the correct place, it was just using the wrong name for
that variable.

Also, clean up and isolate the template that generates an analog
function derivative function definition so it can be
called in a loop (instead of containing a loop), and add a similar
template for generating a declaration for a .h file.

There are still some missing templates for verilog constructs that
ought to be there (case, while, etc.) but I'm keen to move on and
actually start USING these functions, and what I have here is enough
to create all of the derivative functions needed for all of our tricky
models, including MVSG-HV.  So I'm calling a temporary cease fire on
this file, and getting ready to move on to the problem of generating
code to *call* these functions in the main code.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168