Make back-end generate correct code for analog function calls
This commit actually implements the code generation step to handle
analog function calls and their derivatives in normal, non-sacado
context (nothing is changed in sensitivity context).

With this commit, I can process BSIM6.1.1 into C++ that not only
compiles, but runs, and not only runs, but passes both the Xyce test
suite AND the CMC test suite just as well as Xyce 6.11 did (which was
reasonably well, but like most CMC test suites, never perfectly spot
on --- "max rel errors" as high as .2% are common, but are precisely
the same between Xyce 6.11 (sacadoized) and the one I'm
testing (analytic derivs, no sacado).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168