Force instantiation of templated function
The Sacado back-end always forced <AdmsFadType> instantiation of
analog function calls.  When I hacked on the new back end I took that
out, but didn't replace it with a <double>.  That has worked for bsim6
and bsimcmg 107, but breaks bsimcmg 110 because they pass an int into
a function argument sometimes, expecting it to be promoted
automagically to a real.

Reinsert a forced instantiation into analog function calls in normal
context, but <double> this time.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168