Make BSIMCMG 110 using new back-end
Now that the "experimental", de-sacadoized Xyce/ADMS can process
models with analog functions and produce correct derivatives, it can
handle the BSIM-CMG.

The BSIM-CMG 110 model had been manually recoded with analytic
derivatives in the summer of 2018, and then had seven output variables
hacked into it for printing with N() (gm, vth, vgs, and several
others).  The output variables hack now has to be wedged in as a
post-ADMS patch along with the optional "T" node hack.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168