Fix up basic templates for EKV
EKV tripped two previously unnoticed errors in
xyceBasicTemplates_nosac.

The first was that when we're processing voltage limiting stuff, the
"recursiveDetectParameterDependence" template can wind up walking into
a string, and needs to know what to do (nothing) and did not until
now.  I do not know why this didn't bite us with the Sacado version,
but it does in this one.

The second was that while I had added code to correctly differentiate
the "log" function (log base 10), I had not in fact put conditionals
in place so that this code would actually get executed, and instead it
fell through to the "don't know how to differentiate 'log'" error
message.  The EKV model uses "log".

Finally, EKV also uses "$strobe" when "INFO_LEVEL" is set to force
verbosity, and calls functions (like "exp") in the arguments of this
function call.  I had not handled this use case properly, and
Xyce/ADMS was treating it like a function call on the RHS of an
assignment instead of as a bare callfunction.  This means that when I
process strobe, I should be turning off function precomputation (so
the function call is inserted literally instead of being assumed to
have been computed as an intermediate variable earlier).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168