Fix use of "$vt" inside analog functions
$vt(temp) in Verilog-A causes Xyce/ADMS to emit a call to the templated
function "adms_vt" which just takes its argument and multiplies it by
CONSTKoverQ.

When used in an analog function (the templated, return-the-value one),
it should be emitting an instantiation of the function that has the
same typename as the calling templated function.

This one change was all that was needed to get FBH HBT compiled.  The
new Xyce/ADMS back-end no longer requires that FBH HBT's analog
functions have obnoxious and kludgy post-processing hacks applied to
them.  They Just Work so long as this one issue is fixed in the back
end.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168