Fix analog function generation for MVSG-HV
MVSG-HV is the only model we have that actually uses output variables
in analog functions other than the return value.  I had neglected to
write code that emitted required variable declarations for this use
case.

MVSG-HV also has an analog function "mmax" that is supposed to take
two real arguments, but they often call it with "0" (an integer) as
one of its actual args.  Templating the function makes this
problematic when calling that function from inside the non-templated
derivative function "d_mmax", and the compiler hates it.  So this
commit also makes sure to wrap integers in "static_cast<double>" in AF
calls if the analog function is defined with a real argument in that
position.  It was difficult to work out how to do this check based on
the ADMS data tree, and the method I've settled on is somewhat
ridiculous.  I may be able to simplify it.  This version has some
debugging output that I will take out later, when I'm done debugging.

At the same time, I'm removing two files from FBH that are no longer
necessary.  Forgot to add them to an earlier FBH-related commit.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168