Fix stupid logical error
When generating the code for "mapply_binary" when the operator is
multiplication, we're supposed to emit "1.0" when both operands are
identically 1.0.  I had a logic error in the analog function code
that would emit "1.0" if *EITHER* operand was 1.0.  It was a careless
cut and paste operation copying the "if either argument is identically
zero, emit 0.0" code.

The non-analog function version of this template does NOT have the
same error, nor does the division operator.

This is why MVSG-HV was going wrong.  They  have one expression that
has a "*1.0" in one of tthe analog functions, and instead of getting
the whole expression emitted, I was replacing the entire thing with
"1.0".  Oopies.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168