Fix broken BSIM6 issue
On Friday night I unleashed my new Xyce/ADMS on the team, only to find
that it was now failing a number of BSIM6 tests when compiled with GCC
on a few platforms.

I determined late Friday using Valgrind that the issue was an
uninitialized variable error, and Valgrind pointed me right at the
line of code where the uninitialized value was used in a conditional.
The variable being used in the conditional was in fact set in an
assignment immediately prior to the conditional, and there is no way
to make Valgrind tell you *which* upstream quantity was uninitialized,
only which block of memory contains the uninitialized quantity (stack,
heap, etc.).  It was a great, rollicking good time tracking down which
variable was in fact not initialized, all the while fretting over how
in the world my ADMS change could have introduced it.

I did not introduce this problem --- it is present in the original
BSIM6.1.1 code from Berkely, where the local variables "local_sca,"
"local_scb," and "local_scc" are never assigned to unless WPEMOD is
non-zero, but are used on the RHS of assignments *UNCONDITIONALLY*.

The result is undefined behavior that leads to failures that depend on
how you compile, the operating system, and even whether you're doing
serial or parallel run.

The reason we never saw it before is that we were using Sacado for
everything, and the constructor for Sacado objects initializes its
value to zero.  When I switched to analytic/symbolic differentiation
instead of Sacado differentiation, this hidden initialization to zero
went away and exposed the bug in the original BSIM6.1.1 code as it was
written by the BSIM group.

This commit initializes those three variables to zero, eliminating the
reported error.