Comment out unnecessary indentation commands
Way back when, almost all ADMS-generated C++ needed post-generation
patching to get it ready for inclusion in Xyce.  That has not been the
case for a long time.

Back when it was necessary, it was difficult to apply patches
correctly unless the code had been indented according to project style
--- patch got confused by whitespace issues.  Therefore it was routine
to run emacs over every generated file to indent it, then apply
patches.

Very few devices still need patching (and then mostly to insert
capabilities that can't be automatically generated), and the
indentation is generally very, very time consuming.  This makes it
very slow to do work on Xyce/ADMS, because I have to wait forever for
the indentation even though it takes seconds to generate the code
itself.

I am commenting out the indentation commands in all makefiles that
don't apply patches, and changing the "generate_ADMS.sh" script so it
uses a whitespace-ignoring diff when comparing newly generated code to
existing, (presumably) working code.  This speeds my work on the
back-end and should not matter one whit to anyone else.  We don't need
the generated code to look pretty, and if someone *has* to work on a
post-generation C++ file for some bizarre reason, they can always
indent it before working.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168